requirements.txt
yfinance~=0.2.61
lxml
python-telegram-bot~=22.1
python-dotenv~=1.1.0
pandas~=2.2.3